% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seguid.R
\name{seguid}
\alias{seguid}
\alias{useguid}
\alias{lseguid_blunt}
\alias{lseguid_sticky}
\alias{cseguid}
\title{The SEGUID of nucleotide and amino-acid sequences}
\usage{
seguid(seq)

useguid(seq)

lseguid_blunt(seq)

lseguid_sticky(watson, crick, overhang)

cseguid(seq)
}
\arguments{
\item{seq}{A character string.}

\item{watson, crick}{(character string) Two complementary DNA strands.}

\item{overhang}{(integer) Amount of 3' overhang in the 5' side of
the molecule. A molecule with 5' overhang has a negative value.}
}
\value{
A character string.
}
\description{
The SEGUID of nucleotide and amino-acid sequences
}
\examples{
seguid("ACGTACGTACGT")
seguid("acgtACGTacgt")
seguid("ACGTACGTACGT") == seguid("acgtACGTacgt")

useguid("aaa")

lseguid_blunt("ACGGGT")

cseguid("attt")
cseguid("ttta")
cseguid("attt") == cseguid("ttta")

}
\references{
\enumerate{
\item Babnigg G, Giometti CS. A database of unique protein sequence
identifiers for proteome studies. Proteomics.
2006 Aug;6(16):4514-22. \doi{10.1002/pmic.200600032}.
}
}
