% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chksum.R
\name{seguid}
\alias{seguid}
\alias{lsseguid}
\alias{csseguid}
\alias{ldseguid}
\alias{cdseguid}
\title{SEGUID checksum for protein and linear single-stranded DNA}
\usage{
seguid(seq, table = "{DNA}", form = c("long", "short", "both"))

lsseguid(seq, table = "{DNA}", form = c("long", "short", "both"))

csseguid(seq, table = "{DNA}", form = c("long", "short", "both"))

ldseguid(
  watson,
  crick,
  overhang,
  table = "{DNA}",
  form = c("long", "short", "both")
)

cdseguid(watson, crick, table = "{DNA}", form = c("long", "short", "both"))
}
\arguments{
\item{seq}{(character string) The sequence for which the checksum
should be calculated.  The sequence may only comprise of characters
in the alphabet specified by the \code{table} argument.}

\item{table}{(character string) The type of sequence used.
If \code{"{DNA}"} (default), then the input is a DNA sequence.
If \code{"iupac"}, then the input is a DNA sequence specified with
IUPAC ambigous DNA symbols (3).
If \code{"{RNA}"}, then the input is an RNA sequence.
If \code{"{protein}"}, then the input is an amino-acid sequence.}

\item{watson, crick}{(character string) Two reverse-complementary DNA strands.}

\item{overhang}{(integer) Amount of 3' overhang in the 5' side of
the molecule. A molecule with 5' overhang has a negative value.}
}
\value{
\code{seguid()} returns a character string composed of the prefix \verb{seguid-}
followed by a \emph{base64} encoding (2) ("Base 64 Encoding").
A base64 encoding is always 27-character long, and may comprise
non-URL-safe characters.

\code{lsseguid()} returns a character string composed of the prefix \verb{lsseguid-}
followed by a \emph{base64url} encoding (2) ("Base 64 Encoding with URL and
Filename Safe Alphabet").

\code{csseguid()} returns a character string composed of the prefix \verb{csseguid-}
followed by a \emph{base64url} encoding.

\code{cdseguid()} returns a character string composed of the prefix \verb{cdseguid-}
followed by a \emph{base64url} encoding.
}
\description{
SEGUID checksum for protein and linear single-stranded DNA
}
\details{
The Sequence Globally Unique Identifiers (SEGUID) (1) is defined as the
Base64-encoded SHA-1 checksum (4) calculated for the sequence in
uppercase with the trailing pad character \code{=} removed.
In contrast to the original implementation (1), this function returns
the SEGUID checksum prefixed with \verb{seguid-}.
}
\section{Base64 checksums are neither filename nor URL safe}{

The Base64 checksum is not guaranteed to comprise symbols that can
safely be used as-is in Uniform Resource Locator (URL). Specifically,
it may consist of forward slashes (\code{/}) and plus symbols (\code{+}), which
are characters that carry special meaning in a URL.
For the same reason, a Base64 checksum can be guaranteed to be used
as a file or directory name, because it may have a forward slash.
}

\section{Base64url checksums are filename and URL safe}{

The base64url encoding is the base64 encoding with non-URL-safe characters
substituted with URL-safe ones. Specifically, the plus symbol (\code{+}) is
replaced by the minus symbol (\code{-}), and the forward slash (\code{/}) is
replaced by the underscore symbol (\verb{_}).
}

\examples{
## Linear single-stranded DNA:
## GATTACA
seguid("GATTACA")
#> seguid-tp2jzeCM2e3W4yxtrrx09CMKa/8

## Linear single-stranded DNA
## GATTACA
lsseguid("GATTACA")
#> lsseguid-tp2jzeCM2e3W4yxtrrx09CMKa_8

## Circular single-stranded DNA
## GATTACA = ATTACAG = ... = AGATTAC
csseguid("GATTACA")
#> csseguid-mtrvbtuwr6_MoBxvtm4BEpv-jKQ

## Linear double-stranded DNA
## GATTACA
## CTAATGT
seguid::ldseguid("GATTACA", "TGTAATC", overhang = 0)
#> ldseguid-XscjVNyZarYrROVgGXUCleJcMC

## Circular double-stranded DNA
## GATTACA = ATTACAG = ... = AGATTAC
## CTAATGT = TAATGTC = ... = TCTAATG
seguid::cdseguid("GATTACA", "TGTAATC")
#> cdseguid-zCuq031K3_-40pArbl-Y4N9RLnA

}
\references{
\enumerate{
\item Babnigg, G., Giometti, CS. A database of unique protein sequence
identifiers for proteome studies. Proteomics.
2006 Aug;6(16):4514-22. \doi{10.1002/pmic.200600032}.
\item Josefsson, S., The Base16, Base32, and Base64 Data Encodings,
RFC 4648, \doi{10.17487/RFC4648}, October 2006,
\url{https://www.rfc-editor.org/info/rfc4648}.
\item Wikpedia article 'Nucleic acid notation', December 2023.
\url{https://en.wikipedia.org/wiki/Nucleic_acid_notation}.
\item Wikipedia article 'SHA-1' (Secure Hash Algorithm 1), December 2023.
\url{https://en.wikipedia.org/wiki/SHA-1}.
}
}
