% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chksum.R
\name{seguid}
\alias{seguid}
\alias{slseguid}
\alias{dlseguid}
\alias{scseguid}
\alias{dcseguid}
\title{The SEGUID of nucleotide and amino-acid sequences}
\usage{
seguid(seq, table = "dna")

slseguid(seq, table = "dna")

dlseguid(watson, crick, overhang, table = "dna")

scseguid(seq, table = "dna")

dcseguid(watson, crick = rc(watson, table = get_table(table)), table = "dna")
}
\arguments{
\item{seq}{A character string.}

\item{table}{A named character string.}

\item{watson, crick}{(character string) Two complementary DNA strands.}

\item{overhang}{(integer) Amount of 3' overhang in the 5' side of
the molecule. A molecule with 5' overhang has a negative value.}
}
\value{
A character string.
}
\description{
The SEGUID of nucleotide and amino-acid sequences
}
\examples{
seguid("ACGTACGTACGT")

}
\references{
\enumerate{
\item Babnigg G, Giometti CS. A database of unique protein sequence
identifiers for proteome studies. Proteomics.
2006 Aug;6(16):4514-22. \doi{10.1002/pmic.200600032}.
}
}
