% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seguid.R
\name{seguid}
\alias{seguid}
\alias{lseguid}
\alias{useguid}
\alias{cseguid}
\title{The SEGUID of a nucleotide or amino-acid sequence}
\usage{
seguid(seq)

lseguid(seq)

useguid(seq)

cseguid(seq)
}
\arguments{
\item{seq}{A character string.}
}
\value{
A character string.
}
\description{
The SEGUID of a nucleotide or amino-acid sequence
}
\examples{
seguid("ACGTACGTACGT")

useguid("aaa")

cseguid("attt")
cseguid("ttta")
cseguid("attt") == cseguid("ttta")

}
